pyproject.toml
[tool.poetry]
name = "recommendation-engine"
version = "0.1.0"
description = "LLM powered tag based recommendation engine"
authors = ["Vis <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
httpx = "^0.27.0"
numpy = "^1.26.4"
pydantic = "^2.7.1"
google-cloud-firestore = "^2.14.0"
boto3 = "^1.34.89"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1" 